general cleanup of experiments
