+++ .github/workflows/quickstart-test.yaml
@@ -41,4 +41,4 @@ jobs:
       - name: Test quickstart pipeline
         shell: bash -l {0}
         run: |
-          ray job submit --working-dir . -- python experiments/quickstart-synthetic.py --commit_hash exp-${{ github.sha }}
+          ray job submit --working-dir . -- python experiments/quickstart.py --commit_hash exp-${{ github.sha }}
