+++ experiments/164_quality_classifiers.py
@@ -1,3 +1,9 @@
+"""
+Train quality classifiers on different subsets similar to DCLM's classifiers.
+https://github.com/stanford-crfm/marin/issues/164
+TODO: apply these quality classifiers on FineWeb (or DCLM, but that's larger), train models.
+"""
+
 from marin.execution.executor import ExecutorStep, executor_main, this_output_path, versioned
 from marin.processing.classification.fasttext.train_fasttext import (
     DatasetCurationConfig,
