+++ experiments/246_web_extraction_method_training.py
@@ -1,3 +1,8 @@
+"""
+Test different html->text transformation methods (on FineWeb, train 1.4B models).
+https://github.com/stanford-crfm/marin/issues/246
+"""
+
 import logging
 
 from experiments.defaults import default_tokenize, default_train, llama_1_4b_train_config
