+++ experiments/dclm/dclm_data.py
@@ -1,7 +1,3 @@
-from marin.execution.executor import ExecutorStep, executor_main, this_output_path, versioned
-from operations.download.huggingface.download import DownloadConfig, download
-from operations.download.huggingface.download_gated_manual import download_and_upload_to_gcs
-
 """
 Downloads the following datasets:
 - https://huggingface.co/datasets/mlfoundations/dclm-baseline-1.0
@@ -10,6 +6,10 @@ Downloads the following datasets:
 to GCS. The 2nd and 3rd ones above are the other datasets that were used in the DCLM paper (in addition to
     the actual DCLM-Baseline dataset).
 """
+
+from marin.execution.executor import ExecutorStep, executor_main, this_output_path, versioned
+from operations.download.huggingface.download import DownloadConfig, download
+from operations.download.huggingface.download_gated_manual import download_and_upload_to_gcs
 
 ############################################################
 # download DCLM-Baseline dataset
