+++ experiments/dolma/tokenize_dolma.py
@@ -1,14 +1,13 @@
+"""
+Tokenizes the Dolma 1.7 datasets.
+"""
+
 import os.path
 
 from experiments.llama import llama3_tokenizer
 from marin.execution.executor import ExecutorStep, executor_main, this_output_path, versioned
 from marin.processing.tokenize import TokenizeConfig, tokenize
 from marin.processing.tokenize.tokenize import TokenizerStep
-
-"""
-Tokenizes the Dolma 1.7 datasets.
-"""
-
 
 BASE_DIR_DOLMA = "gs://marin-us-central2/raw/dolma/v1.7"
