+++ experiments/llama.py
@@ -4,6 +4,8 @@ Specifies a sequence of Llama 3 models from small to large.
 
 from levanter.models.llama import LlamaConfig
 
+from experiments.defaults import SimpleTrainConfig
+
 llama3_tokenizer = "meta-llama/Meta-Llama-3.1-8B"
 
 llama_150m = LlamaConfig(
@@ -67,6 +69,14 @@ llama_8b = LlamaConfig(
     num_heads=32,
     num_kv_heads=8,
     num_layers=32,
+)
+
+llama_1_4b_train_config = SimpleTrainConfig(
+    tpu_type="v4-128",
+    train_batch_size=1024,
+    num_train_steps=75000,  # 4096 * 1024 * 75000 = 314B tokens
+    learning_rate=3e-4,
+    weight_decay=0.1,
 )
