+++ experiments/quickstart.py
@@ -1,254 +1,234 @@
-# TODO: should delete this file in favor of quickstart-synthetic.py and a
-# defaults.py which has the best practices.
-
-import dataclasses
+import logging
+import os
+from dataclasses import dataclass
 
 import draccus
+from levanter.models.gpt2 import Gpt2Config
+from levanter.trainer import TrainerConfig
 
-import marin.utilities.ray_utils as ray_utils
 from marin.execution.executor import (
+    ExecutorMainConfig,
     ExecutorStep,
     executor_main,
     output_path_of,
     this_output_path,
     versioned,
 )
-
-is_running_locally = ray_utils.is_local_ray_cluster()
-
-USER = "dlwh"
-
-############################################################
-# Download the pretraining data
-from operations.download.huggingface.download import DownloadConfig, download  # noqa
-
-raw_download_step = ExecutorStep(
-    name=f"raw/hello_world_fw-{USER}",
-    fn=download,
-    config=DownloadConfig(
-        hf_dataset_id="skaramcheti/hello_world_fw",
-        revision="8fd6e8e",
-        gcs_output_path=this_output_path(),
-        wait_for_completion=True,
-    ),
-    override_output_path="gs://marin-us-central2/raw/hello_world_fw-pliang-22232d",
-)
-raw_data = output_path_of(
-    raw_download_step, "8fd6e8e/huggingface.co/datasets/skaramcheti/hello_world_fw/resolve/8fd6e8e/data/CC-MAIN-2024-10"
-)
-
-############################################################
-# Transform HTML to text
-
-from marin.schemas.web.convert import ResiliparseConfig, TrafilaturaConfig  # noqa
-from scripts.hello_world_fw.process import FineWebConfig, transform  # noqa
-
-transform_trafilatura_step = ExecutorStep(
-    name=f"documents/hello_world_fw-{USER}-trafilatura",
-    fn=transform,
-    config=FineWebConfig(
-        input_path=raw_data,
-        output_path=this_output_path(),
-        extract_method=versioned("trafilatura"),
-        config=TrafilaturaConfig(
-            favor_precision=versioned(False),
-            favor_recall=versioned(True),
-            include_comments=versioned(False),
-            deduplicate=versioned(False),
-        ),
-    ),
-)
-
-transform_resiliparse_step = ExecutorStep(
-    name=f"documents/hello_world_fw-{USER}-resiliparse",
-    fn=transform,
-    config=FineWebConfig(
-        input_path=raw_data,
-        output_path=this_output_path(),
-        extract_method=versioned("resiliparse"),
-        config=ResiliparseConfig.default_config(),
-    ),
-)
-
-transform_readability_step = ExecutorStep(
-    name=f"documents/hello_world_fw-{USER}-readability",
-    fn=transform,
-    config=FineWebConfig(
-        input_path=raw_data,
-        output_path=this_output_path(),
-        extract_method=versioned("readability"),
-    ),
-)
-
-############################################################
-# Download good data
-
-download_sft_step = ExecutorStep(
-    name="raw/tulu-v2-sft-mixture",
-    fn=download,
-    config=DownloadConfig(
-        hf_dataset_id="allenai/tulu-v2-sft-mixture",
-        revision="6248b17",
-        gcs_output_path=this_output_path(),
-    ),
-)
-
-# TODO: convert the data into Dolma format
-# For now, just use the one that's already given
-sft_data = "gs://marin-us-central2/documents/instruct/v1_olmo_mix/text"
-
-############################################################
-# Train quality classifier
-
-from marin.processing.classification.fasttext.train_fasttext import (  # noqa
+from marin.processing.classification.consolidate import ConsolidateConfig, FilterConfig, consolidate
+from marin.processing.classification.dedupe import DedupeConfig, dedupe
+from marin.processing.classification.fasttext.train_fasttext import (
     DatasetCurationConfig,
     TrainFasttextClassifierConfig,
     train,
 )
+from marin.processing.classification.inference import InferenceConfig, run_inference
+from marin.processing.tokenize import TokenizeConfig, lm_data_config, tokenize
+from marin.schemas.web.convert import HtmlToMarkdownConfig
+from marin.training.training import TrainLmOnPodConfig, run_levanter_train_lm
+from scripts.hello_world_fw.process import FineWebConfig, transform
 
-train_quality_step = ExecutorStep(
-    name=f"classifiers/hello_world_fw-{USER}",
-    fn=train,
-    config=TrainFasttextClassifierConfig(
-        input_doc_paths=[
-            DatasetCurationConfig(
-                input_doc_path=sft_data, label="hq", relative_sampling_rate=0.1, format="dolma_formatted_jsonl"
+logging.basicConfig(level=logging.INFO, format="%(asctime)s - %(levelname)s - %(message)s")
+logger = logging.getLogger(__name__)
+
+
+@dataclass
+class QuickstartExecutorConfig:
+    # all artifacts will be saved in gs://marin-us-central2/{prefix}{commit_hash}
+    commit_hash: str = ""
+    prefix: str = "quickstart-tests"
+
+    # path to synthetic test data
+    synth_data: str = "./tests/quickstart-data"
+
+
+def create_steps(config: QuickstartExecutorConfig) -> list[ExecutorStep]:
+    # ############################################################
+    # Transform HTML to text
+
+    transform_hq_data_step = ExecutorStep(
+        name=os.path.join(config.prefix, config.commit_hash, "hq-transformed"),
+        fn=transform,
+        config=FineWebConfig(
+            input_path=os.path.join(config.synth_data, "pos"),
+            output_path=this_output_path(),
+            extract_method=versioned("readability"),
+            config=HtmlToMarkdownConfig.default_config(),
+        ),
+    )
+
+    transform_lq_data_step = ExecutorStep(
+        name=os.path.join(config.prefix, config.commit_hash, "lq-transformed"),
+        fn=transform,
+        config=FineWebConfig(
+            input_path=os.path.join(config.synth_data, "neg"),
+            output_path=this_output_path(),
+            extract_method=versioned("readability"),
+            config=HtmlToMarkdownConfig.default_config(),
+        ),
+    )
+
+    # ############################################################
+    # Train quality classifier
+
+    train_quality_step = ExecutorStep(
+        name=os.path.join(config.prefix, config.commit_hash, "quality-classifier"),
+        fn=train,
+        config=TrainFasttextClassifierConfig(
+            input_doc_paths=[
+                DatasetCurationConfig(
+                    input_doc_path=output_path_of(transform_hq_data_step),
+                    label="hq",
+                    relative_sampling_rate=1.0,
+                    format="dolma_formatted_jsonl",
+                ),
+                DatasetCurationConfig(
+                    input_doc_path=output_path_of(transform_lq_data_step),
+                    label="lq",
+                    relative_sampling_rate=1.0,
+                    format="dolma_formatted_jsonl",
+                ),
+            ],
+            output_path=this_output_path(),
+            fasttext_args={
+                "lr": 0.001,
+                "minCount": 1,
+                "epoch": 25,
+                "wordNgrams": 2,
+                "dim": 50,
+            },
+        ),
+    )
+
+    ############################################################
+    # Run inference with quality classifier
+
+    inference_hq_step = ExecutorStep(
+        name=os.path.join(config.prefix, config.commit_hash, "hq-inference"),
+        fn=run_inference,
+        config=InferenceConfig(
+            input_path=output_path_of(transform_hq_data_step),
+            output_path=this_output_path(),
+            model_name=output_path_of(train_quality_step),
+            model_type="fasttext",
+            attribute_name="quickstart-fasttext-quality-hq",
+        ),
+    )
+
+    inference_lq_step = ExecutorStep(
+        name=os.path.join(config.prefix, config.commit_hash, "lq-inference"),
+        fn=run_inference,
+        config=InferenceConfig(
+            input_path=output_path_of(transform_lq_data_step),
+            output_path=this_output_path(),
+            model_name=output_path_of(train_quality_step),
+            model_type="fasttext",
+            attribute_name="quickstart-fasttext-quality-lq",
+        ),
+    )
+
+    ############################################################
+    # Deduplicate
+
+    dedupe_step = ExecutorStep(
+        name=os.path.join(config.prefix, config.commit_hash, "dedupe"),
+        fn=dedupe,
+        config=DedupeConfig(
+            input_path=output_path_of(transform_hq_data_step),
+            output_path=this_output_path(),
+        ),
+    )
+
+    ############################################################
+    # Consolidate
+
+    consolidate_step = ExecutorStep(
+        name=os.path.join(config.prefix, config.commit_hash, "consolidate"),
+        fn=consolidate,
+        config=ConsolidateConfig(
+            input_path=output_path_of(transform_hq_data_step),
+            output_path=this_output_path(),
+            filters=[
+                FilterConfig(
+                    type=versioned("classify"),
+                    attribute_path=output_path_of(inference_hq_step),
+                    name=versioned("quickstart-fasttext-quality-hq"),
+                    label="__label__hq",
+                    threshold=versioned(0.1),
+                ),
+                FilterConfig(
+                    type=versioned("remove_spans"),
+                    attribute_path=output_path_of(dedupe_step),
+                    name=versioned("duplicate_text"),
+                ),
+            ],
+        ),
+    )
+
+    ############################################################
+    # Tokenize
+
+    tokenize_step = ExecutorStep(
+        name=os.path.join(config.prefix, config.commit_hash, "tokenized"),
+        fn=tokenize,
+        config=TokenizeConfig(
+            train_paths=output_path_of(consolidate_step),
+            validation_paths=[],
+            cache_path=this_output_path(),
+            tokenizer=versioned("gpt2"),
+        ),
+    )
+
+    ############################################################
+    # Train
+
+    train_step = ExecutorStep(
+        name=os.path.join(config.prefix, config.commit_hash, "train"),
+        fn=run_levanter_train_lm,
+        config=TrainLmOnPodConfig(
+            output_path=this_output_path(),
+            data=lm_data_config(tokenize_step),
+            env={"WANDB_API_KEY": None, "WANDB_MODE": "disabled"},  # Running it locally and turning off wandb
+            tpu_type=None,
+            hf_save_steps=1,
+            model=Gpt2Config(
+                num_layers=2,
+                num_heads=2,
+                seq_len=64,
+                hidden_dim=32,
             ),
-            DatasetCurationConfig(
-                input_doc_path=output_path_of(transform_trafilatura_step),
-                label="lq",
-                relative_sampling_rate=1.0,
-                format="dolma_formatted_jsonl",
-            ),
-        ],
-        output_path=this_output_path(),
-        fasttext_args={"lr": 0.1},
-    ),
-)
+            trainer=TrainerConfig(train_batch_size=1, num_train_steps=2, max_eval_batches=1, require_accelerator=False),
+        ),
+    )
 
-############################################################
-# Run inference with quality classifier
+    return [
+        transform_hq_data_step,
+        transform_lq_data_step,
+        train_quality_step,
+        inference_hq_step,
+        inference_lq_step,
+        dedupe_step,
+        consolidate_step,
+        tokenize_step,
+        train_step,
+    ]
 
-from marin.processing.classification.inference import InferenceConfig, run_inference  # noqa
 
-inference_quality_step = ExecutorStep(
-    name=f"attributes/hello_world_fw-{USER}",
-    fn=run_inference,
-    config=InferenceConfig(
-        input_path=output_path_of(transform_trafilatura_step),
-        output_path=this_output_path(),
-        model_name="allenai/dolma-1_7-fasttext-quality-filter",
-        model_type="fasttext",
-        attribute_name="olmo-fasttext-quality",
-    ),
-)
+@draccus.wrap()
+def main(config: QuickstartExecutorConfig):
+    try:
+        steps = create_steps(config)
+        bucket_prefix = "/tmp"
+        config_executor = ExecutorMainConfig(
+            prefix=bucket_prefix, executor_info_base_path=os.path.join(bucket_prefix, "experiments")
+        )
+        executor_main(config_executor, steps=steps)
+        logger.info(
+            f"Execution completed successfully. All outputs are in {bucket_prefix}/{config.prefix}/{config.commit_hash}"
+        )
+    except Exception as e:
+        logger.error(f"Error in main execution: {e}")
+        raise e
 
-############################################################
-# Deduplicate
-
-from marin.processing.classification.dedupe import DedupeConfig, dedupe  # noqa
-
-dedupe_step = ExecutorStep(
-    name=f"attributes/hello_world_fw-{USER}-dedupe",
-    fn=dedupe,
-    config=DedupeConfig(
-        input_path=output_path_of(transform_trafilatura_step),
-        output_path=this_output_path(),
-    ),
-)
-
-############################################################
-# Consolidate
-
-from marin.processing.classification.consolidate import ConsolidateConfig, FilterConfig, consolidate  # noqa
-
-consolidate_step = ExecutorStep(
-    name=f"documents/hello_world_fw-{USER}-consolidate",
-    fn=consolidate,
-    config=ConsolidateConfig(
-        input_path=output_path_of(transform_trafilatura_step),
-        output_path=this_output_path(),
-        filters=[
-            FilterConfig(
-                type=versioned("classify"),
-                attribute_path=output_path_of(inference_quality_step),
-                name=versioned("olmo-fasttext-quality"),
-                label="__label__hq",
-                threshold=versioned(0.1),
-            ),
-            FilterConfig(
-                type=versioned("remove_spans"),
-                attribute_path=output_path_of(dedupe_step),
-                name=versioned("duplicate_text"),
-            ),
-        ],
-    ),
-)
-
-############################################################
-# Tokenize
-
-from marin.processing.tokenize import TokenizeConfig, tokenize, lm_data_config  # noqa
-
-tokenize_step = ExecutorStep(
-    name=f"tokenized/llama3/hello_world_fw-{USER}",
-    fn=tokenize,
-    config=TokenizeConfig(
-        train_paths=[output_path_of(consolidate_step)],
-        validation_paths=[],
-        cache_path=this_output_path(),
-        tokenizer=versioned("meta-llama/Meta-Llama-3.1-8B"),
-    ),
-)
-
-############################################################
-# Training
-
-from marin.training import TrainLmOnPodConfig, run_levanter_train_lm  # noqa
-
-training_config = draccus.load(TrainLmOnPodConfig, open("config/training/quickstart_run.yaml"))
-
-train_step = ExecutorStep(
-    name=f"checkpoints/quickstart-{USER}",
-    fn=run_levanter_train_lm,
-    config=dataclasses.replace(
-        training_config,
-        output_path=this_output_path(),
-        data=lm_data_config(tokenize_step),
-        tpu_type="v4-8" if not is_running_locally else None,
-    ),
-)
-
-############################################################
-# Evaluate
-
-from marin.evaluation.evaluation_config import EvaluationConfig  # noqa
-from marin.evaluation.run import evaluate  # noqa
-
-evaluate_step = ExecutorStep(
-    name=f"evaluation/hello_world_fw-{USER}",
-    fn=evaluate,
-    config=EvaluationConfig(
-        evaluator="helm",
-        model_name="pf5pe4ut/step-600",
-        # TODO: replace this with `output_path_of(train_step)`
-        model_path="gs://marin-us-central2/checkpoints/quickstart_single_script_docker_test_09_18/pf5pe4ut/hf/pf5pe4ut/step-600",
-        evaluation_path=this_output_path(),
-        evals=["mmlu"],
-    ),
-)
-
-############################################################
 
 if __name__ == "__main__":
-    executor_main(
-        steps=[
-            transform_trafilatura_step,
-            transform_resiliparse_step,  # Not used
-            transform_readability_step,  # Not used
-            # train_quality_step,  # Not used  (TODO: fails right now)
-            tokenize_step,
-            train_step,
-            evaluate_step,
-        ]
-    )
+    main()
